+++ src/lib/components/SecureIframe.js
@@ -11,6 +11,7 @@ export default function SecureIframe(allProps) {
     srcDoc,
     title,
     message,
+    onLoad,
     onMessage,
     iframeResizer,
   } = allProps;
@@ -20,6 +21,7 @@ export default function SecureIframe(allProps) {
   const [loaded, setLoaded] = useState(false);
   const [prevMessage, setPrevMessage] = useState(undefined);
   const ref = React.useRef();
+  const sandbox = "allow-scripts";
 
   const returnIframeResizerProps = () => {
     const result = {
@@ -58,6 +60,12 @@ export default function SecureIframe(allProps) {
       }
     });
 
+    if (allIframeResizerProps.onResized) {
+      result.onResized = ({ height, width }) => {
+        allIframeResizerProps.onResized({ height, width });
+      };
+    }
+
     return result;
   };
 
@@ -70,6 +78,11 @@ export default function SecureIframe(allProps) {
     },
     [ref, onMessage]
   );
+
+  const onLoadHandler = () => {
+    setLoaded(true);
+    onLoad && onLoad();
+  };
 
   useEffect(() => {
     window.addEventListener("message", onMessageEvent, false);
@@ -94,18 +107,13 @@ export default function SecureIframe(allProps) {
       <IframeResizer
         {...returnIframeResizerProps()}
         forwardRef={ref}
-        sandbox="allow-scripts"
-        onLoad={() => setLoaded(true)}
+        sandbox={sandbox}
+        onLoad={onLoadHandler}
       />
     );
   }
 
   return (
-    <iframe
-      {...usedProps}
-      ref={ref}
-      sandbox="allow-scripts"
-      onLoad={() => setLoaded(true)}
-    />
+    <iframe {...usedProps} ref={ref} sandbox={sandbox} onLoad={onLoadHandler} />
   );
 }