+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.3.2
+
+- Added support for `onLoad` event for `<iframe>` tags:
+
+```jsx
+<iframe onLoad={() => { console.log('iframe loaded') }}>
+```
+
+- Added support for `onResized` Iframe Resizer event:
+
+```jsx
+<iframe iframeResizer={{
+  onResized: ({width, height}) => { console.log('iframe resized', width, height) },
+}}>
+```
+
+- Bump `near-api-js` dependency to `^2.1.2`
+
 ## 1.3.1
 
 - Fix the minimum storage deposit for a new account to be attached. This happened because the permission is granted by default, and the logic relied on it first.